BREAKING CHANGE: standardize port number
I [reserved a port number](https://github.com/prometheus/prometheus/wiki/Default-port-allocations) for this exporter. So I switched the inner port from 8000 to this port and switched all the examples.

This allows it to run on the proper port even if it's outside of Docker. Apologies for the breaking change.