Factor out cmdline, allow for scraping build from source or docker.
Created a new envvar MINER_PATH. If it is defined and not empty,
the code will call that in place of the docker module. MINER_PATH
should be the full path to the miner binary, like this:
  MINER_PATH=/opt/miner/bin/miner ./helium_exporter.py

In order to do this, I wrapped the execution in an ExecWrapper
class. It has a run() function that will do either
 docker_container.exec_run()
or
 subprocess.run()

To make dropping this in easier, ExecWrapper.run() has the same
return type as docker_container.exec_run().

Finally, there is some small code reformatting here:
- Reordered the imports because...I couldn't help it

- Whitespace removal in a few places where there either too many
  newlines or just whitespace on an empty line.