Option to gather penalties from all validators
Add an envvar (defaults to off) that will gather penalties for
all validators from the ledger. In an effort to reduce the size
of the collection, this "all" mode will filter out validators
with a total penalty of 0.0.

Regardless of this new envvar, last_heartbeat will still only
be gathered for "this" validator.