Remove full validators scrape; pythonify API.
Clean up the mechanism used for passing parameters to JSONRPC functions.
Do it the Python way, using keyword arguments. Then unpack them into a JSON
dictionary at the appropriate time.

Also, for the time being, don't ask the validator to enumerate all validators
recorded in the ledger. This is an expensive operation and doesn't help
much in the long run. Instead, ask for this validator's stats, only.